+++ pylatex/table.py
@@ -48,7 +48,8 @@ class Tabular(Environment):
         'pos': 'options',
     }
 
-    def __init__(self, table_spec, data=None, pos=None, **kwargs):
+    def __init__(self, table_spec, data=None, pos=None, *, width=None,
+                 **kwargs):
         """
         Args
         ----
@@ -56,13 +57,21 @@ class Tabular(Environment):
             A string that represents how many columns a table should have and
             if it should contain vertical lines and where.
         pos: list
+        width: int
+            The amount of columns that the table has. If this is `None` it is
+            calculated based on the ``table_spec``, but this is only works for
+            simple specs. In cases where this calculation is wrong override the
+            width using this argument.
 
         References
         ----------
         * https://en.wikibooks.org/wiki/LaTeX/Tables#The_tabular_environment
         """
 
-        self.width = _get_table_width(table_spec)
+        if width is None:
+            self.width = _get_table_width(table_spec)
+        else:
+            self.width = width
 
         super().__init__(data=data, options=pos,
                          arguments=table_spec, **kwargs)